+++ CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[3.1.3](https://github.com/TerraDharitri/drt-js-sdk-web-wallet-cross-window-provider/pull/91)] - 2025-07-10
+
+- [Added safeguard for multiple logout requests](https://github.com/TerraDharitri/drt-js-sdk-web-wallet-cross-window-provider/pull/90)
+
+## [[3.1.2](https://github.com/TerraDharitri/drt-js-sdk-web-wallet-cross-window-provider/pull/89)] - 2025-07-09
+
+- [Update PopupConsent to be inside WindowManager](https://github.com/TerraDharitri/drt-js-sdk-web-wallet-cross-window-provider/pull/88)
+
 ## [[3.1.1](https://github.com/TerraDharitri/drt-js-sdk-web-wallet-cross-window-provider/pull/87)] - 2025-06-12
 
 - [Added session storage support to preserve session ID across window refreshes](https://github.com/TerraDharitri/drt-js-sdk-web-wallet-cross-window-provider/pull/86)
 - [Added `finalizeResetStateRequest` string optional payload](https://github.com/TerraDharitri/drt-js-sdk-web-wallet-cross-window-provider/pull/85)
-
 
 ## [[3.1.0](https://github.com/TerraDharitri/drt-js-sdk-web-wallet-cross-window-provider/pull/84)] - 2025-05-06